Restructure accesskey commands, add strings
