+++ cmd/idp-ldap-accesskey-info.go
@@ -1,0 +1,193 @@
+// Copyright (c) 2015-2023 MinIO, Inc.
+//
+// This file is part of MinIO Object Storage stack
+//
+// This program is free software: you can redistribute it and/or modify
+// it under the terms of the GNU Affero General Public License as published by
+// the Free Software Foundation, either version 3 of the License, or
+// (at your option) any later version.
+//
+// This program is distributed in the hope that it will be useful
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+// GNU Affero General Public License for more details.
+//
+// You should have received a copy of the GNU Affero General Public License
+// along with this program.  If not, see <http://www.gnu.org/licenses/>.
+
+package cmd
+
+import (
+	"strings"
+	"time"
+
+	"github.com/charmbracelet/lipgloss"
+	humanize "github.com/dustin/go-humanize"
+	"github.com/fatih/color"
+	"github.com/minio/cli"
+	json "github.com/minio/colorjson"
+	"github.com/minio/mc/pkg/probe"
+	"github.com/minio/pkg/v2/console"
+)
+
+var idpLdapAccesskeyInfoCmd = cli.Command{
+	Name:         "info",
+	Usage:        "info about given access key pairs for LDAP",
+	Action:       mainIDPLdapAccesskeyInfo,
+	Before:       setGlobalsFromContext,
+	Flags:        append(idpLdapPolicyAttachFlags, globalFlags...),
+	OnUsageError: onUsageError,
+	CustomHelpTemplate: `NAME:
+  {{.HelpName}} - {{.Usage}}
+
+USAGE:
+  {{.HelpName}} [FLAGS] TARGET ACCESSKEY [ACCESSKEY...]
+
+FLAGS:
+  {{range .VisibleFlags}}{{.}}
+  {{end}}
+EXAMPLES:
+  TODO: add examples
+	`,
+}
+
+type ldapAccesskeyMessage struct {
+	op            string
+	Status        string          `json:"status"`
+	AccessKey     string          `json:"accessKey"`
+	SecretKey     string          `json:"secretKey,omitempty"`
+	ParentUser    string          `json:"parentUser,omitempty"`
+	AccountStatus string          `json:"accountStatus,omitempty"`
+	ImpliedPolicy bool            `json:"impliedPolicy,omitempty"`
+	Policy        json.RawMessage `json:"policy,omitempty"`
+	Name          string          `json:"name,omitempty"`
+	Description   string          `json:"description,omitempty"`
+	Expiration    *time.Time      `json:"expiration,omitempty"`
+}
+
+func (m ldapAccesskeyMessage) String() string {
+	switch m.op {
+	case "info":
+		// t := newPrettyRecord(2,
+		// 	Row{"AccessKey", "AccessKey"},
+		// 	Row{"ParentUser", "ParentUser"},
+		// 	Row{"Policy", "Policy"},
+		// 	Row{"Name", "Name"},
+		// 	Row{"Description", "Description"},
+		// 	Row{"Expiration", "Expiration"},
+		// )
+		// expirationStr := "NONE"
+		// if m.Expiration != nil && !m.Expiration.IsZero() && !m.Expiration.Equal(timeSentinel) {
+		// 	expirationStr = humanize.Time(*m.Expiration)
+		// }
+		// policyStr := "embedded"
+		// if m.ImpliedPolicy {
+		// 	policyStr = "implied"
+		// }
+		// return t.buildRecord(m.AccessKey, m.ParentUser, policyStr, m.Name, m.Description, expirationStr)
+
+		expirationStr := "NONE"
+		if m.Expiration != nil && !m.Expiration.IsZero() && !m.Expiration.Equal(timeSentinel) {
+			expirationStr = humanize.Time(*m.Expiration)
+		}
+		policyStr := "embedded"
+		if m.ImpliedPolicy {
+			policyStr = "implied"
+		}
+
+		labelStyle := lipgloss.NewStyle().Foreground(lipgloss.Color("#04B575")) // green
+		o := strings.Builder{}
+
+		o.WriteString(iFmt(0, "%s %s\n", labelStyle.Render("Access Key:"), m.AccessKey))
+		o.WriteString(iFmt(0, "%s %s\n", labelStyle.Render("Parent User:"), m.ParentUser))
+		o.WriteString(iFmt(0, "%s %s\n", labelStyle.Render("Policy:"), policyStr))
+		o.WriteString(iFmt(0, "%s %s\n", labelStyle.Render("Name:"), m.Name))
+		o.WriteString(iFmt(0, "%s %s\n", labelStyle.Render("Description:"), m.Description))
+		o.WriteString(iFmt(0, "%s %s\n\n", labelStyle.Render("Expiration:"), expirationStr))
+
+		return o.String()
+
+	case "create":
+		// t := newPrettyRecord(2,
+		// 	Row{"Title", "Title"},
+		// 	Row{"AccessKey", "AccessKey"},
+		// 	Row{"SecretKey", "SecretKey"},
+		// 	Row{"Expiration", "Expiration"},
+		// )
+		// titleStr := "Added new access key for LDAP user " + m.ParentUser
+		// expirationStr := "NONE"
+		// if m.Expiration != nil && !m.Expiration.IsZero() && !m.Expiration.Equal(timeSentinel) {
+		// 	expirationStr = fmt.Sprintf("%s", m.Expiration)
+		// }
+		// return t.buildRecord(titleStr, m.AccessKey, m.SecretKey, expirationStr)
+
+		expirationStr := "NONE"
+		if m.Expiration != nil && !m.Expiration.IsZero() && !m.Expiration.Equal(timeSentinel) {
+			expirationStr = m.Expiration.String()
+		}
+
+		labelStyle := lipgloss.NewStyle().Foreground(lipgloss.Color("#04B575")) // green
+		o := strings.Builder{}
+
+		o.WriteString(iFmt(0, "%s %s\n", labelStyle.Render("User DN:   "), m.ParentUser))
+		o.WriteString(iFmt(0, "%s %s\n", labelStyle.Render("Access Key:"), m.AccessKey))
+		o.WriteString(iFmt(0, "%s %s\n", labelStyle.Render("Secret Key:"), m.SecretKey))
+		o.WriteString(iFmt(0, "%s %s\n\n", labelStyle.Render("Expiration:"), expirationStr))
+
+		return o.String()
+	case "delete":
+		return console.Colorize("DeleteAccessKey", "Successfully deleted access key "+m.AccessKey)
+	}
+	return ""
+}
+
+func (m ldapAccesskeyMessage) JSON() string {
+	jsonMessageBytes, e := json.MarshalIndent(m, "", " ")
+	fatalIf(probe.NewError(e), "Unable to marshal into JSON.")
+
+	return string(jsonMessageBytes)
+}
+
+func mainIDPLdapAccesskeyInfo(ctx *cli.Context) error {
+	if len(ctx.Args()) < 2 {
+		showCommandHelpAndExit(ctx, 1) // last argument is exit code
+	}
+
+	console.SetColor("AccessKey", color.New(color.FgCyan, color.Bold))
+	console.SetColor("ParentUser", color.New(color.FgCyan))
+	console.SetColor("Expiration", color.New(color.FgYellow))
+	console.SetColor("Policy", color.New(color.FgGreen))
+	console.SetColor("Name", color.New(color.FgGreen))
+
+	args := ctx.Args()
+	aliasedURL := args.Get(0)
+	accessKeys := args.Tail()
+
+	// Create a new MinIO Admin Client
+	client, err := newAdminClient(aliasedURL)
+	fatalIf(err, "Unable to initialize admin connection.")
+
+	for _, accessKey := range accessKeys {
+		res, e := client.InfoServiceAccount(globalContext, accessKey)
+		if e != nil {
+			errorIf(probe.NewError(e), "Unable to retrieve access key "+accessKey+" info.")
+		} else {
+			m := ldapAccesskeyMessage{
+				op:            "info",
+				AccessKey:     accessKey,
+				Status:        "success",
+				ParentUser:    res.ParentUser,
+				AccountStatus: res.AccountStatus,
+				ImpliedPolicy: res.ImpliedPolicy,
+				Policy:        json.RawMessage(res.Policy),
+				Name:          res.Name,
+				Description:   res.Description,
+				Expiration:    res.Expiration,
+			}
+
+			printMsg(m)
+		}
+	}
+
+	return nil
+}
